Light creation/deletion sockets done, deploy config finished
I centralized all the socket communication to the server for creation of
light beams. A user creates and then is notified when the light is
addded to the database.

Configuration for nodejitsu was added as environment variables to
conceal private information. I also forgot to add sails-mongo as a
dependency which is why none of my db connections were working
initially.